refactor command organisation
